+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]()
+
+### Added
+
+### Changed
+
+### Fixed
+- `BottleRocketPage` - Fixes attack history being limited.
+- `TimeControl` - Fixes reset time function counting 1 minute too short.
+
 ## [1.0.0](https://raw.githubusercontent.com/Synogun/FarmRPGPlus/refs/tags/1.0.0/dist/FarmRPGPlus.user.js)
 
 ### Added